Day4_Calling_LLM_Python/requirements.txt
python-dotenv==1.0.1
groq==0.13.1
streamlit==1.37.1
rich==13.7.1
requests==2.32.3
tenacity==9.0.0
